updated request function to expose the response on error
