+++ sdk/src/utils/request.ts
@@ -1,4 +1,4 @@
-import type { ResponseTransformer } from "../index.js";
+import { extractData } from './extract-data.js';
 
 /**
  * Request helper providing default settings
@@ -10,49 +10,20 @@ import type { ResponseTransformer } from "../index.js";
  */
 export const request = async <Output = any>(
 	url: string,
-	options: RequestInit,
-	formatter?: ResponseTransformer<Output> | null
+	options: RequestInit
 ): Promise<Output> => {
 	options.headers =
 		typeof options.headers === 'object' && !Array.isArray(options.headers)
 			? (options.headers as Record<string, string>)
 			: {};
 
-	const defaultFormatter = (data: Output | { data: Output }) => {
-		if (typeof data === 'object' && data && 'data' in data) {
-			return data.data;
-		}
+	const response = await globalThis.fetch(url, options).catch((reason) => {
+		throw { reason, response };
+	});
 
-		return data;
-	};
+	const data = await extractData(response).catch((reason) => {
+		throw { reason, response };
+	});
 
-	const outputFormatter = formatter !== undefined && formatter !== null ? formatter : defaultFormatter;
-
-	const response = await globalThis.fetch(url, options);
-
-	const data = await extractData(response)
-		.catch((reason) => {
-			throw { reason, response };
-		})
-
-	return outputFormatter(data, options);
+	return data;
 };
-
-async function extractData(response: Response) {
-	const type = response.headers.get('Content-Type')?.toLowerCase();
-
-	if (type?.startsWith('application/json')) {
-		const result = await response.json();
-		if (!response.ok) throw result;
-		return result;
-	}
-
-	if (type?.startsWith('text/html') || type?.startsWith('text/plain')) {
-		const result = await response.text();
-		if (!response.ok) throw result;
-		return result;
-	}
-
-	// empty body fallback
-	return;
-}
